Merge admin reporting feature into main branch

+++ client/src/pages-admin/Dashboard/DashboardStats.jsx
@@ -14,6 +14,11 @@ import {
 import { validateAdminResponse } from "../../lib/response.js";
 import ErrorComponent from "../../components/ErrorComponent.jsx";
 
+/**
+ * Component used to fetch and render database stats
+ * @returns {JSX.Element}
+ */
+
 function DashboardStats() {
   const COLORS = ["#0088FE", "#00C49F", "#FFBB28", "#FF8042", "#185c36"];
   const { apihost, auth, setAuth } = useOutletContext();
@@ -117,7 +122,7 @@ function DashboardStats() {
                     <div>
                       {row.total_interactions} (
                       {(
-                        (parseFloat(row.total_interactions) /
+                        (parseInt(row.total_interactions) /
                           parseInt(stats.count_total)) *
                         100
                       ).toFixed()}
